implement snapshot option in cloud_backup
